requirements.txt
google-genai
pyautogui